extract custom node to another file, rename type to 'step' instead of 'special'
